Clean up duplicated types + move type defs into sensible places
